requirements2.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
authlib==1.6.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
cyclopts==3.22.5
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
docutils==0.22
duckduckgo-search==8.1.1
email-validator==2.2.0
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.11.3
filelock==3.19.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lazy-object-proxy==1.11.0
litellm==1.75.7
lxml==6.0.0
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.5
mcp==1.13.0
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.6.4
numpy==2.3.2
openai==1.99.9
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-httpx==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
pandas==2.3.1
parse==1.20.2
pathable==0.4.4
pdfkit==1.0.0
pillow==11.3.0
pluggy==1.6.0
primp==0.15.0
propcache==0.3.2
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
pypdf2==3.0.1
pyperclip==1.9.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
starlette==0.47.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.1
wkhtmltopdf==0.2
wrapt==1.17.3
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
